Syntax Is Corresponding To Eachother
The reason the assets were not showing up was because i had one with a Capital 'A' vs lower cased 'a' for the folder Assets. With is with the github repo could not locate its contents despite them being in the same folder. Attention to the little things are truly key!!